fix: Extract Context into own module
